Moved integration api calls
